Backend Code Marathon (Incompleto)/CodeMarathon-main/Backend/requirements.txt
django >= 4.0, < 4.1
djangorestframework >= 3.13.1, < 3.14.0
django-cors-headers >= 3.11.0, < 3.12.0
docker >= 4.2.2, < 4.3.0
# To load data
PyYAML >= 6.0, < 6.1
# For testing purposes
ddt >= 1.4.4, < 1.5.0
# For deployment
gunicorn >= 20.1.0, < 20.2.0
